[People] Move test data to sub-dir
